重构feedFragment，抽离基类，添加刷新以及加载更多

+++ app/src/main/java/cn/imrhj/cowlevel/network/service/CowLevel.kt
@@ -15,5 +15,5 @@ interface CowLevel {
      * 获取feed时间线
      */
     @GET("/api/feed/timeline")
-    fun feedTimeline(@Query("last_id") id: Long = 0): Observable<ApiModel<FeedApiModel>>
+    fun feedTimeline(@Query("last_id") id: Int = 0): Observable<ApiModel<FeedApiModel>>
 }
